moving function getToken to auth
